requirements.txt
langgraph==0.2.71
python-dotenv==1.0.1
langchain-anthropic==0.3.7
langchain-ollama==0.2.3
langchain-openai==0.3.5
langchain==0.3.18
ipython==8.32.0
fastapi==0.115.8
uvicorn==0.34.0
streamlit==1.42.0